Van hoang feature3

+++ TechForum/src/main/java/com/example/techforum/service/blog/BlogService.java
@@ -3,7 +3,9 @@ package com.example.techforum.service.blog;
 import com.example.techforum.dto.BlogDto;
 import com.example.techforum.dto.BlogDtoNew;
 import com.example.techforum.model.Blogs;
+import com.example.techforum.model.Categories;
 import com.example.techforum.model.Users;
+import com.example.techforum.repository.ICategoryRepo;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,9 +22,11 @@ import java.util.stream.Collectors;
 public class BlogService implements IBlogService{
     @Autowired
     private IBlogRepo iBlogRepository;
+    @Autowired
+    private ICategoryRepo iCategoryRepo;
 
     @Override
-    public Blogs updateBlog(Long id,@Valid BlogDto updatedBlog) {
+    public Blogs updateBlog(Integer id,@Valid BlogDto updatedBlog) {
         if (iBlogRepository.existsById(id)) {
             Blogs blog = iBlogRepository.findById(id).get();
             blog.setCategory(updatedBlog.getCategory());
@@ -37,7 +41,7 @@ public class BlogService implements IBlogService{
     }
 
     @Override
-    public void acctiveBlog(Long id) {
+    public void acctiveBlog(Integer id) {
         Blogs blog = iBlogRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Blog with id " + id + " not found"));
         blog.setStatus(true);
         iBlogRepository.save(blog);
@@ -63,7 +67,6 @@ public class BlogService implements IBlogService{
                 .collect(Collectors.toList());
     }
 
-
     @Override
     public Blogs addNewBlog(BlogDto blog) {
         return iBlogRepository.save(dtoToObject(blog));
@@ -72,7 +75,7 @@ public class BlogService implements IBlogService{
 
 
     @Override
-    public BlogDto findOne(long id) {
+    public BlogDto findOne(Integer id) {
         Blogs blog = iBlogRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Blog not found with id: " + id));
 
@@ -81,8 +84,12 @@ public class BlogService implements IBlogService{
 
 
     @Override
-    public void delete(long id) {
-        iBlogRepository.deleteById(id);
+    public void delete(Integer id) {
+        if (iBlogRepository.existsById(id)){
+            iBlogRepository.deleteById(id);
+        }else {
+            throw new EntityNotFoundException("Blog not found with id: " + id);
+        }
     }
 
 
@@ -100,8 +107,13 @@ public class BlogService implements IBlogService{
 
 
     @Override
-    public Blogs findById(long id) {
+    public Blogs findById(Integer id) {
         return iBlogRepository.findObject(id);
+    }
+    @Override
+    public List<BlogDto> findByCategoryId(Integer id) {
+        Categories categories = iCategoryRepo.findById(id).get();
+        return iBlogRepository.findAllByCategory(categories);
     }
 
     public Blogs dtoToObject(BlogDto blogDTO){
